requirements.txt
bokeh=1.0.4
python>3.5
